Refactored Entry Edit screen.
The Confirm field is now used to determine if the password actually
matches and prevents the Save button from being enabled if it's not.
Only problem is it only updates the CanSave status when tabbing out of
the password/confirm fields. WinRT is missing UpdateSourceTrigger!

Also, went ahead and made Save navigate back to the previous screen
(probably what the user will usually want). And fixed the screen so that
it doesn't update the underlying database unless the user actually taps
the Save button.