Different style for root folder in main view.
The root folder will now show in a different style on the main grid view
so that it stands out from the subfolders which are also shown as top
level groups in the grid. Now just need to pick a better color. :-)

Also fixed the typo in the "Coverters" namespace to "Converters".