Handle navigating to non-existent folder.
This can happen when the user deletes a folder but then uses the Back
button to go back to the screen where they deleted it from. Since the
UUID of that folder is no longer in the database, we needed to show a
message saying that the folder could not be found instead of crashing.
This is now working.